// src/app/dtos/usuario.dto.ts
export class UsuarioDto {
  idUsuario: string = '';
  nombreUsuario: string = '';
  apellidoPaternoUsuario: string = '';
  apellidoMaternoUsuario: string = '';
  dniUsuario: string = '';
  claveUsuario: string = '';
  nivelUsuario: string = '';
  correoUsuario: string = '';
  celularUsuario: string = '';
  direccionUsuario: string = '';
  turnoUsuario: string = '';
  estadoUsuario: string = '';
}